+++ R/create_yaml_header.R
@@ -1,9 +1,9 @@
 #' Convert list objects into R Markdown YAML headers
 #'
-#' This function tweaks the behavior of \code{\link[yaml]{as.yaml}} to return a string
-#' which can immediately be used as an R Markdown YAML header. It's designed
-#' to accept both deeply nested lists and simpler list formats to make
-#' reasoning about your header easier.
+#' This function tweaks the behavior of \code{\link[yaml]{as.yaml}} to return
+#' a string which can immediately be used as an R Markdown YAML header.
+#' It's designed to accept both deeply nested lists and simpler list formats
+#' to make reasoning about your header easier.
 #'
 #' @param ... A set of objects that will be combined into the YAML header.
 #' Objects may be provided as lists (the structure
@@ -11,7 +11,7 @@
 #' or as single-item named vectors (passing "title" = "My Report" to ... will
 #' translate to title: "My Report").
 #'
-#' @param line.sep,indent,omap,column.major,unicode,precision,indent.mapping.sequence,handlers
+#' @param line.sep,indent,unicode,indent.mapping.sequence,handlers
 #' Additional arguments to be passed to \code{\link[yaml]{as.yaml}}
 #'
 #' @return Returns a string formatted for use as an R Markdown YAML header.
@@ -41,10 +41,7 @@
 create_yaml_header <- function(...,
                                line.sep = c("\n", "\r\n", "\r"),
                                indent = 2,
-                               omap = FALSE,
-                               column.major = TRUE,
                                unicode = TRUE,
-                               precision = getOption("digits"),
                                indent.mapping.sequence = FALSE,
                                handlers = NULL) {
   yaml.parts <- list(...)
@@ -73,10 +70,10 @@ create_yaml_header <- function(...,
       header.content,
       line.sep,
       indent,
-      omap,
-      column.major,
+      omap = FALSE,
+      column.major = TRUE,
       unicode,
-      precision,
+      precision = getOption("digits"),
       indent.mapping.sequence,
       handlers
     ),