+++ R/import_pattern.R
@@ -1,8 +1,8 @@
 #' Quickly import plaintext files.
 #'
 #' Longer, multi-chunk patterns can benefit from being developed in files
-#' independent of the rest of a draft. This is a quick wrapper function to import
-#' those patterns as objects for assembly into a draft.
+#' independent of the rest of a draft. This is a quick wrapper function to
+#' import those patterns as objects for assembly into a draft.
 #'
 #' @param filepath A valid character string to the plaintext file containing
 #' the pattern.
@@ -12,11 +12,9 @@
 #' @family import functions
 #'
 #' @examples
-#' \dontrun{
-#' export_template("my sample pattern", out.Rmd)
-#' import_template("out.Rmd")
-#' }
-#'
+#' pattern_file <- tempfile("out", tempdir(), ".Rmd")
+#' export_template("my sample pattern", pattern_file)
+#' import_pattern(pattern_file)
 #' @export
 
 import_pattern <- function(filepath) {