+++ prepare_development_branch.sh
@@ -5,7 +5,6 @@ R -e "rmarkdown::render('README.Rmd')"
 R -e "devtools::build_manual()"
 R -e "devtools::build_site()"
 
-
 R -e "styler::style_dir()"
 R -e 'styler::style_dir(filetype = "Rmd")'